refactor: Auth flow for twitch -> privy -> creators app
