+++ apps/main-landing/src/app/creators/page.tsx
@@ -9,10 +9,22 @@ import { OAuthCallbackHandler } from './components/oauth-callback-handler';
 export default function Landing() {
   const searchParameters = useSearchParams();
   const customToken = searchParameters.get('token');
+  const name = searchParameters.get('name');
+  const displayName = searchParameters.get('displayName');
+  const pfp = searchParameters.get('pfp');
+  const email = searchParameters.get('email');
+  const login = searchParameters.get('login');
 
   return (
     <div className="relative flex h-screen">
-      <OAuthCallbackHandler authToken={customToken} />
+      <OAuthCallbackHandler
+        authToken={customToken}
+        name={name}
+        displayName={displayName}
+        pfp={pfp}
+        email={email}
+        login={login}
+      />
       <ScrollArea
         type="always"
         customScrollEventName="creatorsLandingPageScroll"
