+++ apps/main-landing/src/app/creators/providers.tsx
@@ -1,8 +1,6 @@
-'use client';
 import NiceModal from '@ebay/nice-modal-react';
 import { WithPortal } from '@idriss-xyz/ui/providers/with-portal';
-import { ReactNode, useCallback, useEffect, useState } from 'react';
-import { PrivyProvider } from '@privy-io/react-auth';
+import { ReactNode } from 'react';
 import { TooltipProvider } from '@idriss-xyz/ui/tooltip';
 
 import { QueryProvider } from '@/providers';
@@ -12,45 +10,11 @@ type Properties = {
 };
 
 export const Providers = ({ children }: Properties) => {
-  const [token, setToken] = useState<string | null>(null);
-  const [loadingToken, setLoadingToken] = useState(true);
-
-  useEffect(() => {
-    const t = localStorage.getItem('custom-auth-token');
-    setToken(t);
-    setLoadingToken(false);
-  }, []);
-
-  const getCustomAuthToken = useCallback(() => {
-    return Promise.resolve(token ?? undefined);
-  }, [token]);
-
   return (
     <QueryProvider>
       <WithPortal>
         <NiceModal.Provider>
-          <TooltipProvider delayDuration={400}>
-            <PrivyProvider
-              appId={process.env.NEXT_PUBLIC_PRIVY_APP_ID ?? ''}
-              config={{
-                appearance: {
-                  theme: 'light',
-                  accentColor: '#000000',
-                  showWalletLoginFirst: true,
-                },
-                // This handles the custom Twitch login flow
-                customAuth: {
-                  getCustomAccessToken: getCustomAuthToken,
-                  isLoading: loadingToken,
-                },
-                embeddedWallets: {
-                  createOnLogin: 'users-without-wallets',
-                },
-              }}
-            >
-              {children}
-            </PrivyProvider>
-          </TooltipProvider>
+          <TooltipProvider delayDuration={400}>{children}</TooltipProvider>
         </NiceModal.Provider>
       </WithPortal>
     </QueryProvider>
